// src/app/model/OpeningHoursTo.ts
export class OpeningHoursTo {
    mondayOpeningHour: string | undefined;
    mondayClosingHour: string | undefined;
    tuesdayOpeningHour: string | undefined;
    tuesdayClosingHour: string | undefined;
    wednesdayOpeningHour: string | undefined;
    wednesdayClosingHour: string | undefined;
    thursdayOpeningHour: string | undefined;
    thursdayClosingHour: string | undefined;
    fridayOpeningHour: string | undefined;
    fridayClosingHour: string | undefined;
    saturdayOpeningHour: string | undefined;
    saturdayClosingHour: string | undefined;
    sundayOpeningHour: string | undefined;
    sundayClosingHour: string | undefined;
}